separate code that works with shell or terminal
