+++ src/kubernetes/kubernetesTools.ts
@@ -5,8 +5,8 @@ import safesh from 'shell-escape-tag';
 import { AzureConstants } from '../azure/azureTools';
 import { Errorable, failed, succeeded } from '../types/errorable';
 import { globalState, telemetry } from '../extension';
-import { output } from '../output';
-import { shellCodeError } from '../shell';
+import { output } from '../shell/output';
+import { shellCodeError } from '../shell/shell';
 import { TelemetryErrorEventNames } from '../telemetry';
 import { parseJson } from '../utils/jsonUtils';
 import { ContextTypes, setVSCodeContext } from '../vscodeContext';
